Remove unnecessary class and modify logout redirection url.
