+++ unihub/users-service/src/main/webapp/WEB-INF/shiro.ini
@@ -2,6 +2,7 @@
 listener = org.apache.shiro.config.event.LoggingBeanEventListener
 
 shiro.loginUrl = "/rest/users/login"
+shiro.logoutUrl = "/rest/users/logout"
 # todo: set to true
 shiro.postOnlyLogout = true
 
@@ -29,11 +30,13 @@ passwordMatcher.passwordService = $passwordService
 jdbcRealm.credentialsMatcher = $passwordMatcher
 
 authc.loginUrl = /login
+logout.redirectUrl = /defaultLogoutUrl
 user.loginUrl = /login
 
 [urls]
 /rest/users/add = anon
 /rest/users/login = anon
+/rest/users/logout = logout
 /rest/users/nbUsers = anon
 /rest/users/** = authc
